Added freezable module
Necessary to ensure objects are frozen after being initialized. This module is specialized for being *prepended* in order to ensure proper freezing of an object.

Reek has been updated to ignore module initialization which, under normal circumstances, would be valid but this module is designed for prepending and needs this behavior to work.