Added diffability
Necessary to allow wholable objects to not only compare themslves but also know what is different. This implementation is a direct copy from the link:https://alchemists.io/projects/refinements[Refinements] gem in order to not introduce a gem dependency but this might need to change in the future if this duplication gets unwieldy.

Milestone: minor