Added pattern matching
Necessary to support pattern matching so these objects can have the same behavior as found with other value objects like `Struct` and `Data`. A new `#define_class_methods` method was introduced to reduce the strain on the `#included` method.

Reek was updated to ignore the utility functions since they are a necessary evil in this case.

Milestone: minor